Renaming Membership -> Cluster
Cluster brings the notion of a distributed system much more than Membership. Also update the module names to have cluster, suuchi-membership-* -> suuchi-clustre-*
