+++ suuchi-cluster-scalecube/pom.xml
@@ -1,0 +1,40 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <artifactId>suuchi-cluster-scalecube</artifactId>
+    <version>0.2.2-SNAPSHOT</version>
+    <name>Suuchi Scalecube Membership</name>
+    <description>Suuchi Scalecube based Membership</description>
+    <packaging>jar</packaging>
+
+    <parent>
+        <groupId>in.ashwanthkumar</groupId>
+        <artifactId>suuchi</artifactId>
+        <version>0.2.2-SNAPSHOT</version>
+    </parent>
+
+    <properties>
+        <log4j.version>2.6.2</log4j.version>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>in.ashwanthkumar</groupId>
+            <artifactId>suuchi-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>io.scalecube</groupId>
+            <artifactId>scalecube-cluster</artifactId>
+            <version>0.9.0</version>
+        </dependency>
+        <dependency>
+            <groupId>io.reactivex</groupId>
+            <artifactId>rxscala_${scala.lib.version}</artifactId>
+            <version>0.26.2</version>
+        </dependency>
+
+    </dependencies>
+
+</project>
