--- suuchi-membership-atomix/pom.xml
@@ -1,52 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>suuchi-membership-atomix</artifactId>
-    <version>0.2.2-SNAPSHOT</version>
-    <name>Suuchi Atomix Membership</name>
-    <description>Suuchi Atomix based Membership</description>
-    <packaging>jar</packaging>
-
-    <parent>
-        <groupId>in.ashwanthkumar</groupId>
-        <artifactId>suuchi</artifactId>
-        <version>0.2.2-SNAPSHOT</version>
-    </parent>
-
-    <properties>
-        <log4j.version>2.6.2</log4j.version>
-        <atomix.version>1.0.0-rc9</atomix.version>
-        <catalyst.version>1.1.1</catalyst.version>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>in.ashwanthkumar</groupId>
-            <artifactId>suuchi-core</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>io.atomix</groupId>
-            <artifactId>atomix</artifactId>
-            <version>${atomix.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.atomix</groupId>
-            <artifactId>atomix-resource</artifactId>
-            <version>${atomix.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.atomix.catalyst</groupId>
-            <artifactId>catalyst-transport</artifactId>
-            <version>${catalyst.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>io.atomix.catalyst</groupId>
-            <artifactId>catalyst-netty</artifactId>
-            <version>${catalyst.version}</version>
-        </dependency>
-
-    </dependencies>
-
-</project>
